Give preference to placing items into bins with a lower copy count
In general items will go into the smallest space large enough to receive them.  However, now they will be placed into a space in a bin with a lower copy count, even if there is a smaller space that would receive them in a bin with a larger copy count.